[Refactor]: 재사용성을 위해 웹 schema 분리
- 기존 inference_router에 있던 구조를 따로 분리
- 추후 diagnosis router에서도 사용
