+++ backend/src/routers/schemas/__init__.py
@@ -1,0 +1,1 @@
+from .items import InferenceInput, ResultItem, InferenceResult
